Move all zx scripts to typescript
