+++ scripts/tsconfig.json
@@ -1,11 +1,11 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
-    "allowJs": true,
-    "checkJs": true,
     "noEmit": true,
     "resolveJsonModule": true,
-    "types": ["zx", "node"]
+    "types": ["node"],
+    "composite": true,
+    "target": "es6"
   },
-  "include": ["*.mjs", "**/*.mjs"]
+  "include": ["*.ts", "**/*.ts"]
 }
